testing/web-platform/tests/html/semantics/scripting-1/the-script-element/script-before-after-events.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>script beforescriptexecute/afterscriptexecute events</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#the-script-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var t1 = async_test("'beforescriptexecute'/'afterscriptexecute' events have been fired"),
      a = false,
      b = false;

  var before = function(e) {
    b = true;
    test(function(){
      assert_true(e.isTrusted);
      assert_true(e.bubbles);
      assert_true(e.cancelable);
    }, "'beforescriptexecute' event is trusted, bubbles and is cancelable");
  };

  var after = function(e) {
    a = true;
    test(function(){
      assert_true(e.isTrusted);
      assert_true(e.bubbles);
      assert_false(e.cancelable);
    }, "'afterscriptexecute' event is trusted, bubbles and isn't cancelable");
  };

  document.body.onload = t1.step_func_done(function(){
    assert_true(a && b);
  });
</script>
<script onbeforescriptexecute=before(event) onafterscriptexecute=after(event)>
  document.querySelector("script");
</script>
